chore: move updated buttons to `tailwind` directory
